Google Analyticsの改善 (#62)
* fix google analytics

* delete _root.tsx

* format pages.gen.ts

* fix google analytics
