+++ src/pages/_layout.tsx
@@ -4,18 +4,22 @@ import '../styles.css';
 
 import { Header } from '../components/header';
 import { Footer } from '../components/footer';
+import { GoogleAnalytics } from '../components/google-analytics';
 
 type RootLayoutProps = { children: ReactNode };
 
 export default async function RootLayout({ children }: RootLayoutProps) {
   return (
-    <div className="overflow-hidden font-['Nunito']">
-      <Header />
-      <main className="m-6 flex justify-center *:min-h-64 *:min-w-64 lg:my-0 lg:min-h-svh lg:pb-20">
-        {children}
-      </main>
-      <Footer />
-    </div>
+    <>
+      <GoogleAnalytics />
+      <div className="overflow-hidden font-['Nunito'] has-[#sp-menu[data-expanded='true']]:overflow-hidden">
+        <Header />
+        <main className="m-6 flex justify-center *:min-h-64 *:min-w-64 lg:my-0 lg:min-h-svh lg:pb-20">
+          {children}
+        </main>
+        <Footer />
+      </div>
+    </>
   );
 }
